Refactor continued
