+++ foolscap/meta_data/__init__.py
@@ -1,0 +1,6 @@
+from foolscap.meta_data.io import load_meta
+
+from foolscap.meta_data.common import note_exists
+from foolscap.meta_data.common import new_component
+from foolscap.meta_data.common import update_component
+from foolscap.meta_data.common import upgrade_components
