+++ foolscap/note_content.py
@@ -1,16 +1,23 @@
 import os
 
-from file_paths import NOTE_FOLDERS # io handler
-from parse_text import (
-    load_text, # IO handler
-    edit_text # IO handler,
-    unique_heading, # new component?
-    shift_lines, # move here
+from foolscap.file_paths import NOTE_FOLDERS
+from foolscap.handle_note_io import (
+    load_text,
+    edit_text,
+    save_text,
+    remove_text,
+    unique_text,
 )
-from meta_data import (
+
+from foolscap.meta_data import (
+    note_exists,
     new_component,
     update_component,
+    upgrade_components,
 )
+
+def update_notes():
+    upgrade_components()
 
 
 def save_note(new_note, temp_file=False):
@@ -21,16 +28,14 @@ def save_note(new_note, temp_file=False):
     """
     # Used for parsing '.txt' notes.
     if not temp_file:
-        new_note = load_text(new_note)
-    titles, contents  = new_component(new_note)
+        new_note = load_text(new_note, new_note=True)
+    titles, contents = new_component(new_note)
     if temp_file:
-        title = note_titles[0]
-        content = contents[0]
-        save_text(title, content)
-        print("\n\tSaved note: '{}'.\n".format(title))
+        # save_text(note_titles[0], contents[0])
+        print("\n\tSaved note: '{}'.\n".format(titles[0]))
     else:
         for title, content in zip(titles, contents):
-            save_text(title, content)
+            # save_text(title, content)
             print("\n\tAdded: '{}'.\n".format(title))
 
 
@@ -72,7 +77,7 @@ def delete_note(note):
         delete_file = folders['GET_NOTE'].format(note_name=note)
 
         # Move unique heading to components
-        recycle_bin = unique_heading(note, folder='IN_BIN')
+        recycle_bin = unique_text(note, folder='IN_BIN')
         bin_note = folders['BIN_NOTE'].format(note_name=recycle_bin)
 
         os.rename(delete_file, bin_note)
@@ -121,4 +126,31 @@ def move_lines(note):
         edit_text(editing=edited_note)
 
         shift_lines(from_note, note)
+
+
+def remove_moving_lines(note):
+    _lines = [line for line in note if line[:1] != '>']
+    return _lines
+
+
+def shift_lines(name_from_note, name_to_note):
+    # load text, return note, delete, save new
+    take_from_note = load_text(path_from)
+    remove_text(path_from)
+
+    replace_note = get_contents(take_from_note)[0]
+    replace_note = remove_moving_lines(replace_note)
+    save_text(name_from_note, replace_note)
+
+    # load text, return note, apply new, delete old, save new
+    apply_to_note = load_text(name_to_note)
+    # Get contents with be tricky to refactor
+    apply_to_note = get_contents(apply_to_note)[0]
+
+    line_index = len(apply_to_note) - 2
+    # Insert moved lines into other note.
+    apply_to_note[line_index:line_index] = get_moving_lines(take_from_note)
+    remove_text(name_to_note)
+    save_text(name_to_note, apply_to_note)
+
